Re-upload instead of cancelling when a Shipmondo order is deleted
Deleting a sales order in Shipmondo no longer cancels the Sylius order. The
old CancelOrderHandler did two contradictory things on `orders/delete` — it
cancelled the Sylius order *and* reset the upload state for re-upload — but a
cancelled order is never re-upload-eligible (the provider requires state=new),
so the reset was dead weight and the order just ended up cancelled.

Replace it with ResetUploadStateHandler: on `orders/delete` it resets the
order's Shipmondo upload state (shipmondoState -> pending) and clears its
shipmondoId, so the next `upload-orders` run re-uploads it. Cancellation is a
Sylius-side decision and is intentionally not driven by a Shipmondo deletion.

Drops the order state-machine dependency from the handler; updates the service
wiring, the test, and UPGRADE.md.
