+++ src/Webhook/Handler/ResetUploadStateHandler.php
@@ -11,24 +11,22 @@ use Setono\Shipmondo\Enum\WebhookResourceName;
 use Setono\SyliusShipmondoPlugin\Webhook\OrderResolverInterface;
 use Setono\SyliusShipmondoPlugin\Webhook\RemoteEvent;
 use Setono\SyliusShipmondoPlugin\Workflow\OrderWorkflow;
-use Sylius\Abstraction\StateMachine\StateMachineInterface;
-use Sylius\Component\Order\OrderTransitions;
 use Symfony\Component\Workflow\WorkflowInterface;
 
 /**
- * Reacts to a sales order being deleted in Shipmondo: cancels the Sylius order and resets its Shipmondo
- * upload state (the order no longer exists in Shipmondo).
+ * Reacts to a sales order being deleted in Shipmondo: the order no longer exists there, so its
+ * Shipmondo upload state is reset to pending and its Shipmondo id is cleared, which makes the next
+ * `upload-orders` run re-upload it.
  *
- * Shipmondo has no dedicated "cancel" action for sales orders — cancelling an order archives/deletes
- * it and fires the `orders/delete` webhook (the deleted order reports `order_status: "archived"`).
+ * Deleting a sales order in Shipmondo is intentionally *not* treated as a cancellation — cancelling an
+ * order is a Sylius-side decision.
  */
-final class CancelOrderHandler implements RemoteEventHandlerInterface
+final class ResetUploadStateHandler implements RemoteEventHandlerInterface
 {
     use ORMTrait;
 
     public function __construct(
         private readonly OrderResolverInterface $orderResolver,
-        private readonly StateMachineInterface $stateMachine,
         private readonly WorkflowInterface $orderWorkflow,
         ManagerRegistry $managerRegistry,
     ) {
@@ -46,22 +44,12 @@ final class CancelOrderHandler implements RemoteEventHandlerInterface
             return;
         }
 
-        $changed = false;
-
-        if ($this->stateMachine->can($order, OrderTransitions::GRAPH, OrderTransitions::TRANSITION_CANCEL)) {
-            $this->stateMachine->apply($order, OrderTransitions::GRAPH, OrderTransitions::TRANSITION_CANCEL);
-            $changed = true;
-        }
-
-        // the order no longer exists in Shipmondo, so reset its upload state back to pending
-        if ($this->orderWorkflow->can($order, OrderWorkflow::TRANSITION_RESET)) {
-            $this->orderWorkflow->apply($order, OrderWorkflow::TRANSITION_RESET);
-            $changed = true;
-        }
-
-        if (!$changed) {
+        if (!$this->orderWorkflow->can($order, OrderWorkflow::TRANSITION_RESET)) {
             return;
         }
+
+        $this->orderWorkflow->apply($order, OrderWorkflow::TRANSITION_RESET);
+        $order->setShipmondoId(null);
 
         $this->getManager($order)->flush();
     }
